requirements.txt
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
entsoe-py==0.7.8
dash>=2.14.0
plotly>=5.17.0
gunicorn>=21.2.0